Create Dataset-ch4-p3
In the question set, questions 1 and 14 are taken from Imam et al. (2012); questions 2 to 4 are taken from Turner et al. (2010). Questions 5 to 13 and 26 to 32 are from the first team of domain experts from the University of Melbourne; questions 15 to 25 are from the second team of domain experts from Monash University.
As the second resource, questions were gathered from literature published by high standard research teams and credible publishers, as these types of questions are of interest to neuroscientists; important; up-to-date; and worth studying. However, after creating the list, all questions were validated by an expert.